requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dicttoxml==1.7.4
Flask==0.12.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
futures==3.1.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
MySQL-python==1.2.5
python-envcfg==0.2.0
redis==2.10.5
requests==2.18.2
requests-futures==0.9.7
SQLAlchemy==1.1.12
urllib3==1.22
Werkzeug==0.12.2
Flask-login==0.4.0